Fix: [ui] fix for column title not showing two lines
Column title suppose to shoe account type and number in two separate lines. But
the way we pass the title text and render it was HTML encoding `<br>` tags and
showing them as text.

This fix the issue by allowing us to send `|` (pipe) symbol as indicators where
the lines should break.